+++ domain/src/main/java/com/zerobase/cms/domain/DomainApplication.java
@@ -1,0 +1,12 @@
+package com.zerobase.cms.domain;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = "com.zerobase.cms")
+public class DomainApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DomainApplication.class, args);
+    }
+}
