+++ domain/src/main/java/com/zerobase/cms/domain/domain/common/UserVo.java
@@ -1,0 +1,13 @@
+package com.zerobase.cms.domain.domain.common;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class UserVo {
+
+    private Long id;
+    private String email;
+
+}
