+++ user-api/build.gradle
@@ -29,6 +29,8 @@ ext{
 
 
 dependencies {
+    tasks.register("prepareKotlinBuildScriptModel"){}
+    implementation project(':domain')
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
@@ -41,7 +43,6 @@ dependencies {
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-//    runtimeOnly 'mysql:mysql-connector-java'
     implementation 'mysql:mysql-connector-java:8.0.33'
 
     //baseEntity
@@ -51,7 +52,6 @@ dependencies {
     
     //랜덤코드 라이브러리
     implementation 'org.apache.commons:commons-lang3:3.12.0'
-
 }
