--- user-api/src/main/java/com/example/user/config/FeignConfig.java
@@ -1,16 +1,0 @@
-package com.example.user.config;
-
-import feign.auth.BasicAuthRequestInterceptor;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FeignConfig {
-
-    @Qualifier(value = "mailgun")
-    @Bean
-    public BasicAuthRequestInterceptor basicAuthRequestInterceptor() {
-        return new BasicAuthRequestInterceptor("api", "21a31457bc83066a6a9797c8b27f125b-e71583bb-a47c6c10");
-    }
-}
