--- user-api/src/main/java/com/example/user/controller/SignUpController.java
@@ -1,29 +1,0 @@
-package com.example.user.controller;
-
-import com.example.user.application.SignUpApplication;
-import com.example.user.domain.SignUpForm;
-import feign.Response;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/signup")
-@RequiredArgsConstructor
-public class SignUpController {
-
-    private final SignUpApplication signUpApplication;
-
-
-    @PostMapping
-    public ResponseEntity<String> customerSignUp(@RequestBody SignUpForm form) {
-        return ResponseEntity.ok(signUpApplication.customerSignUp(form));
-    }
-
-    @GetMapping("/verify/customer")
-    public ResponseEntity<String> verifyCustomer(@RequestParam String email, @RequestParam String code){
-        signUpApplication.customerVerify(email,code);
-        return ResponseEntity.ok("인증이 완료되었습니다.");
-    }
-}
