--- user-api/src/main/java/com/example/user/domain/model/Customer.java
@@ -1,47 +1,0 @@
-package com.example.user.domain.model;
-
-import com.example.user.domain.SignUpForm;
-import com.fasterxml.jackson.databind.ser.Serializers;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.envers.AuditOverride;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.Locale;
-
-@Entity
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@AuditOverride(forClass = BaseEntity.class)
-public class Customer extends BaseEntity {
-    @Id
-    @Column(name = "id", nullable = false)
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-//    @Column(unique = true)
-    private String email;
-    private String name;
-    private String password;
-    private String phone;
-    private LocalDate birth;
-
-    private LocalDateTime verifyExpiredAt;
-    private String verificationCode;
-    private boolean verify;
-
-    public static Customer from(SignUpForm form){
-        return Customer.builder()
-                .email(form.getEmail().toLowerCase(Locale.ROOT))
-                .password(form.getPassword())
-                .name(form.getName())
-                .birth(form.getBirth())
-                .phone(form.getPhone())
-                .verify(false)
-                .build();
-    }
-}
