--- user-api/src/main/java/com/example/user/exception/CustomerException.java
@@ -1,15 +1,0 @@
-package com.example.user.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomerException extends RuntimeException {
-    private final ErrorCode errorCode;
-
-    public CustomerException(ErrorCode errorCode){
-        super(errorCode.getDetail());
-        this.errorCode = errorCode;
-    }
-
-
-}
