--- user-api/src/main/java/com/example/user/exception/ExceptionController.java
@@ -1,34 +1,0 @@
-package com.example.user.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-@Slf4j
-public class ExceptionController {
-
-    @ExceptionHandler({
-            CustomerException.class
-    })
-    public ResponseEntity<ExceptionResponse> customRequestException(final CustomerException c) {
-        log.warn("api Exception : {}", c.getErrorCode());
-        return ResponseEntity.badRequest().body(new ExceptionResponse(c.getMessage(),c.getErrorCode()));
-    }
-
-
-    @Getter
-    @ToString
-    @AllArgsConstructor
-    public static class ExceptionResponse {
-
-    private String message;
-    private ErrorCode errorCode;
-}
-}
