--- user-api/src/main/java/com/example/user/service/SignUpCustomerService.java
@@ -1,64 +1,0 @@
-package com.example.user.service;
-
-import com.example.user.domain.SignUpForm;
-import com.example.user.domain.model.Customer;
-import com.example.user.domain.repository.CustomerRepository;
-import com.example.user.exception.CustomerException;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
-import java.util.Locale;
-import java.util.Optional;
-
-import static com.example.user.exception.ErrorCode.*;
-
-@Service
-@RequiredArgsConstructor
-public class SignUpCustomerService {
-
-    private final CustomerRepository customerRepository;
-
-
-    public Customer signUp(SignUpForm form){
-        return customerRepository.save(Customer.from(form));
-    }
-
-    public boolean isEmailExist(String email){
-        return customerRepository.findByEmail(email.toLowerCase(Locale.ROOT))
-                .isPresent();
-    }
-
-    @Transactional
-    public void verifyEmail(String email, String code){
-        Customer customer = customerRepository.findByEmail(email)
-                .orElseThrow(()->new CustomerException(NOT_FOUND_USER));
-        if(customer.isVerify()){
-            throw new CustomerException(ALREADY_VERIFY);
-        }
-        if(!customer.getVerificationCode().equals(code)){
-            throw new CustomerException(WRONG_VERIFICATION);
-        }
-        if(customer.getVerifyExpiredAt().isBefore(LocalDateTime.now())){
-            throw new CustomerException(EXPIRE_CODE);
-        }
-        customer.setVerify(true);
-    }
-
-    @Transactional
-    public LocalDateTime ChangeCustomerValidateEmail(Long customerId, String verificationCode){
-        Optional<Customer> customerOptional = customerRepository.findById(customerId);
-
-        if(customerOptional.isPresent()){
-            Customer customer = customerOptional.get();
-            customer.setVerificationCode(verificationCode);
-            customer.setVerifyExpiredAt(LocalDateTime.now().plusDays(1));
-            return customer.getVerifyExpiredAt();
-        }else{
-
-            return LocalDateTime.now();
-        }
-
-    }
-}
