+++ user-api/src/main/java/com/zerobase/cms/user/config/JwtConfig.java
@@ -1,0 +1,17 @@
+package com.zerobase.cms.user.config;
+
+import com.zerobase.cms.domain.config.JwtAuthenticationProvider;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class JwtConfig {
+
+    private static final String SECRET_KEY = "ThisIsAReallyLongSecretKeyThatIsAtLeast64BytesLongForHS512Algorithm123456";
+
+    @Bean
+    public JwtAuthenticationProvider jwtAuthenticationProvider() {
+        return new JwtAuthenticationProvider(SECRET_KEY);
+    }
+}
+
