+++ user-api/src/main/java/com/zerobase/cms/user/controller/CustomerController.java
@@ -1,0 +1,42 @@
+package com.zerobase.cms.user.controller;
+
+import com.zerobase.cms.domain.config.JwtAuthenticationProvider;
+import com.zerobase.cms.domain.domain.common.UserVo;
+import com.zerobase.cms.user.domain.customer.CustomerDto;
+import com.zerobase.cms.user.domain.model.Customer;
+import com.zerobase.cms.user.domain.repository.CustomerRepository;
+import com.zerobase.cms.user.exception.CustomerException;
+import com.zerobase.cms.user.service.CustomerService;
+import feign.Response;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import static com.zerobase.cms.user.exception.ErrorCode.NOT_FOUND_USER;
+
+@RestController
+@RequestMapping("/customer")
+@RequiredArgsConstructor
+public class CustomerController {
+
+    private final JwtAuthenticationProvider provider;
+    private final CustomerService service;
+    private final CustomerService customerService;
+
+    @GetMapping("/getInfo")
+    public ResponseEntity<CustomerDto> getInfo(@Parameter(in = ParameterIn.HEADER, description = "Access Token", required = true)
+                                                   @RequestHeader(name="X-AUTH-TOKEN") String token){
+        UserVo vo = provider.getUserVo(token);
+        Customer c = customerService.findByIdAndEmail(vo.getId(),vo.getEmail()).orElseThrow(
+                ()->new CustomerException(NOT_FOUND_USER)
+        );
+        return ResponseEntity.ok(CustomerDto.form(c));
+    }
+
+
+}
