+++ user-api/src/main/java/com/zerobase/cms/user/controller/SignInController.java
@@ -1,0 +1,24 @@
+package com.zerobase.cms.user.controller;
+
+import com.zerobase.cms.user.application.SignInApplication;
+import com.zerobase.cms.user.domain.SignInForm;
+import com.zerobase.cms.user.domain.model.Customer;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/singIn")
+@RequiredArgsConstructor
+public class SignInController {
+
+    private final SignInApplication signInApplication;
+
+    @PostMapping("/customer")
+    public ResponseEntity<String> signInCustomer(@RequestBody SignInForm form){
+        return ResponseEntity.ok(signInApplication.customerLoginToken(form));
+    }
+}
