+++ user-api/src/main/java/com/zerobase/cms/user/domain/SignInForm.java
@@ -1,0 +1,9 @@
+package com.zerobase.cms.user.domain;
+
+import lombok.Getter;
+
+@Getter
+public class SignInForm {
+    private String email;
+    private String password;
+}
