+++ user-api/src/main/java/com/zerobase/cms/user/domain/customer/CustomerDto.java
@@ -1,0 +1,19 @@
+package com.zerobase.cms.user.domain.customer;
+
+import com.zerobase.cms.user.domain.model.Customer;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+@AllArgsConstructor
+public class CustomerDto {
+
+    private Long id;
+    private String email;
+
+    public static CustomerDto form(Customer customer){
+      return new CustomerDto(customer.getId(), customer.getEmail());
+    }
+}
