+++ user-api/src/main/java/com/zerobase/cms/user/domain/model/BaseEntity.java
@@ -1,0 +1,20 @@
+package com.zerobase.cms.user.domain.model;
+
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+import java.time.LocalDateTime;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(value = {AuditingEntityListener.class})
+public abstract class BaseEntity {
+
+    @CreatedBy
+    private LocalDateTime createdAt;
+    @LastModifiedBy
+    private LocalDateTime modifiedAt;
+}
