+++ user-api/src/main/java/com/zerobase/cms/user/domain/repository/CustomerRepository.java
@@ -1,0 +1,16 @@
+package com.zerobase.cms.user.domain.repository;
+
+import com.zerobase.cms.user.domain.model.Customer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+
+    Optional<Customer> findByEmail(String email);
+
+
+
+}
