+++ user-api/src/main/java/com/zerobase/cms/user/filter/CustomerFilter.java
@@ -1,0 +1,34 @@
+package com.zerobase.cms.user.filter;
+
+import com.zerobase.cms.domain.config.JwtAuthenticationProvider;
+import com.zerobase.cms.domain.domain.common.UserVo;
+import com.zerobase.cms.user.service.CustomerService;
+import jakarta.servlet.*;
+import jakarta.servlet.annotation.WebFilter;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+
+import java.io.IOException;
+
+@WebFilter(urlPatterns = "/customer/*")
+@RequiredArgsConstructor
+public class CustomerFilter implements Filter {
+
+    private final JwtAuthenticationProvider jwtAuthenticationProvider;
+    private final CustomerService customerService;
+
+
+    @Override
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
+        HttpServletRequest req = (HttpServletRequest) request;
+        String token = req.getHeader("X-AUTH-TOKEN");
+        if(!jwtAuthenticationProvider.validateToken(token)) {
+            throw new ServletException("Invalid token");
+        }
+        UserVo vo = jwtAuthenticationProvider.getUserVo(token);
+        customerService.findByIdAndEmail(vo.getId(),vo.getEmail()).orElseThrow(
+                ()->new ServletException("Invalid token")
+        );
+        chain.doFilter(request, response);
+    }
+}
